mapping service initiated in main activity
