+++ app/src/main/java/com/scepticalphysiologist/dmaple/ui/Recorder.kt
@@ -2,10 +2,13 @@ package com.scepticalphysiologist.dmaple.ui
 
 import android.annotation.SuppressLint
 import android.content.pm.PackageManager
+import android.os.Bundle
 import android.text.format.DateUtils
 import android.view.MotionEvent
+import android.view.View
 import androidx.core.content.ContextCompat.checkSelfPermission
 import androidx.lifecycle.ViewModelProvider
+import com.scepticalphysiologist.dmaple.MainActivity
 import com.scepticalphysiologist.dmaple.R
 import com.scepticalphysiologist.dmaple.databinding.RecorderBinding
 import com.scepticalphysiologist.dmaple.etc.PermissionSets
@@ -31,21 +34,14 @@ class Recorder : DMapLEPage<RecorderBinding>(RecorderBinding::inflate) {
         // Keep the screen on, so that the camera stays on.
         //binding.root.keepScreenOn = true
 
-        // Once the view is inflated, connect the mapping service, passing it the camera preview.
+        // Once the view is inflated.
         binding.root.post {
-            model.connectMappingService(binding.root.context)
+            MainActivity.setCameraPreviewSurface(binding.cameraAndRoi.getCameraPreview())
+            binding.cameraAndRoi.setExposureSlider(0.5f)
+            binding.cameraAndRoi.setSavedRois(model.getMappingRois())
+            setUIState()
         }
 
-        // Once the mapping service is connected: set it's preview; restore any saved ROIs from it;
-        // and set the UI state.
-        model.mappingServiceConnected.observe(viewLifecycleOwner) { isConnected ->
-            if(isConnected) {
-                model.setCameraPreviewSurface(binding.cameraAndRoi.getCameraPreview())
-                binding.cameraAndRoi.setExposureSlider(0.5f)
-                binding.cameraAndRoi.setSavedRois(model.getMappingRois())
-                setUIState()
-            }
-        }
 
         // Whenever the saved ROIs have changed in the view, transfer these to the mapping
         // service (via the view model) so they are persisted.
